% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_hr_synths}
\alias{acs_hr_synths}
\title{American Community Survey higher-risk synthetic data}
\format{
\subsection{\code{acs_hr_synths}}{

A list of 30 data frames with 1,000 rows and 11 columns:
\describe{
\item{county}{fct, county}
\item{gq}{fct, group quarter kind}
\item{sex}{fct, sex}
\item{marst}{fct, marital status}
\item{hcovany}{fct, health insurance status}
\item{empstat}{fct, employment status}
\item{classwkr}{fct, employment kind (ex: self-employed, etc.)}
\item{age}{dbl, age (in years)}
\item{famsize}{dbl, household/family size}
\item{transit_time}{dbl, transit time to work (in minutes)}
\item{inctot}{dbl, annual income}
}
}
}
\source{
\url{https://usa.ipums.org/usa/}
}
\usage{
acs_hr_synths
}
\description{
A list of 30 samples of higher-risk synthetic data derived from \code{acs_conf},
generated using weakly variable resampling and models overfit to the confidential data.
}
\details{
Original data source:
Steven Ruggles, Sarah Flood, Matthew Sobek, Daniel Backman, Annie Chen,
Grace Cooper, Stephanie Richards, Renae Rogers, and Megan Schouweiler.
IPUMS USA: Version 15.0 [dataset]. Minneapolis, MN: IPUMS, 2024.
https://doi.org/10.18128/D010.V15.0
}
\keyword{datasets}
