% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_specks.R
\name{add_specks}
\alias{add_specks}
\title{Add SPECKS to discrimination object}
\usage{
add_specks(discrimination, split = TRUE, group_by_q = NULL)
}
\arguments{
\item{discrimination}{A discrimination with added propensities}

\item{split}{A logical for if the metric should be calculated separately for
the training/testing split. Defaults to TRUE.}

\item{group_by_q}{An optional grouping variable to calculate SPECKS within each group.}
}
\value{
A discrimination with SPECKS. When split is TRUE or when group_by_q is
provided, the SPECKS will be calculated within each group and the returned table
contains a ".group" and/or ".sample" column as appropriate.
}
\description{
Add SPECKS to discrimination object
}
\seealso{
Other Utility metrics: 
\code{\link{add_pmse}()},
\code{\link{add_pmse_ratio}()},
\code{\link{add_propensities}()},
\code{\link{add_propensities_tuned}()},
\code{\link{compute_ci_overlap}()},
\code{\link{discrimination}()},
\code{\link{util_ci_overlap}()},
\code{\link{util_co_occurrence}()},
\code{\link{util_ks_distance}()},
\code{\link{util_proportions}()}
}
\concept{Utility metrics}
