% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_qid_mi.R
\name{aggregate_qid_eval}
\alias{aggregate_qid_eval}
\title{Aggregate and count factor factor-level quasi-identifiers for eval_data}
\usage{
aggregate_qid_eval(eval_data, keys)
}
\arguments{
\item{eval_data}{An \code{eval_data} object.}

\item{keys}{A character vector of column quasi-identifiers}
}
\value{
A data.frame of results
}
\description{
Aggregate and count factor factor-level quasi-identifiers for eval_data
}
