% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birthwt}
\alias{birthwt}
\title{Example birthweight data}
\format{
An object of class \code{data.frame} with 47 rows and 12 columns.
}
\source{
\url{https://www.sheffield.ac.uk/mash/statistics/datasets}
}
\usage{
birthwt
}
\description{
This dataset contains information on newborn babies and their parents.
}
\details{
\describe{
\item{length}{Length of baby (cm)}
\item{birthweight}{Weight of baby (kg)}
\item{headcirumference}{Head Circumference}
\item{gestation}{Gestation (weeks)}
\item{motherage}{Maternal age}
\item{mnocig}{Number of cigarettes smoked per day by mother}
\item{mheight}{Mothers height (cm)}
\item{mppwt}{Mothers pre-pregnancy weight (kg)}
\item{fage}{Father's age}
\item{fedyrs}{Father’s years in education}
\item{fnocig}{Number of cigarettes smoked per day by father}
\item{fweight}{Father's weight (kg)}
}
}
\seealso{
\link{birthwt_ind}, \link{birthwt_dep}
}
\keyword{datasets}
