% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birthwt_dep}
\alias{birthwt_dep}
\title{Dependent synthetic birthweight data}
\format{
An object of class \code{data.frame} with 47 rows and 12 columns.
}
\usage{
birthwt_dep
}
\description{
Synthetic version of birthwt data, with column relationships
Generated with synthpop, using method = CART
}
\seealso{
\link{birthwt}, \link{birthwt_ind}
}
\keyword{datasets}
