% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birthwt_ind}
\alias{birthwt_ind}
\title{Independent synthetic birthweight data}
\format{
An object of class \code{data.frame} with 47 rows and 12 columns.
}
\usage{
birthwt_ind
}
\description{
Synthetic version of birthwt data, without column relationships.
Generated with synthpop, using method = parametric and a diagonal
predictor matrix.
}
\seealso{
\link{birthwt}, \link{birthwt_dep}
}
\keyword{datasets}
