% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_df_compatibility}
\alias{check_df_compatibility}
\title{check that df_syn and df_ori are compatible}
\usage{
check_df_compatibility(df_syn, df_orig)
}
\arguments{
\item{df_syn}{synthesized data frame}

\item{df_orig}{original data frame}
}
\description{
syn_df is a data frame with
}
\keyword{internal}
