% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_occurrence.R
\name{co_occurrence}
\alias{co_occurrence}
\title{Construct a co-occurrence matrix}
\usage{
co_occurrence(data, na.rm = FALSE)
}
\arguments{
\item{data}{A tibble with numeric variables}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\value{
A co-occurrence matrix
}
\description{
Construct a co-occurrence matrix
}
