% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_na_helper.R
\name{convert_na_to_level}
\alias{convert_na_to_level}
\title{Convert \code{NA} values to \code{"NA"} for categorical variables}
\usage{
convert_na_to_level(data)

convert_na_to_level(data)
}
\arguments{
\item{data}{A data frame or tibble}
}
\value{
A data frame or tibble with \code{NA} converted to \code{"NA"}

A data frame or tibble with \code{NA} converted to \code{"NA"}
}
\description{
Convert \code{NA} values to \code{"NA"} for categorical variables

Convert \code{NA} values to \code{"NA"} for categorical variables
}
