% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_fit.R
\name{cor_fit}
\alias{cor_fit}
\title{Calculate the correlation fit metric of a confidential data set.}
\usage{
cor_fit(postsynth, data)
}
\arguments{
\item{postsynth}{A postsynth object from tidysynthesis or a tibble}

\item{data}{an original (observed) data set.}
}
\value{
A \code{list} of fit metrics:
\itemize{
\item \code{correlation_original}: correlation matrix of the original data.
\item \code{correlation_synthetic}: correlation matrix of the synthetic data.
\item \code{correlation_difference}: difference between \code{correlation_synthetic} and
\code{correlation_original}.
\item \code{correlation_fit}: square root of the sum of squared differences between
\code{correlation_synthetic} and \code{correlation_original}, divided by the number of
cells in the correlation matrix.
}
}
\description{
Calculate the correlation fit metric of a confidential data set.
}
\seealso{
Other utility functions: 
\code{\link{moments}()},
\code{\link{percentiles}()},
\code{\link{tails}()}
}
\concept{utility functions}
