% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_combined_df}
\alias{create_combined_df}
\title{create syn_df}
\usage{
create_combined_df(df_syn, df_orig)
}
\arguments{
\item{df_syn}{synthesized data frame}

\item{df_orig}{original data frame}
}
\description{
syn_df is a data frame with
}
\keyword{internal}
