% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plots.R
\name{create_na_matrix_plot}
\alias{create_na_matrix_plot}
\title{Create a black/white missingness matrix plot for a data set.}
\usage{
create_na_matrix_plot(data, na_values = NULL)
}
\arguments{
\item{data}{A data.frame}

\item{na_values}{An optional scalar or vector of values (in addition to \code{NA})
that should be treated as missing}
}
\value{
A \code{ggplot2} plot
}
\description{
Create a black/white missingness matrix plot for a data set.
}
