% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_baseline.R
\name{disc_baseline}
\alias{disc_baseline}
\title{Compute baseline disclosure risk metrics using confidential data.}
\usage{
disc_baseline(
  eval_data,
  qid_keys = NULL,
  sens_keys = NULL,
  tclose_metric_cat = c("linf", "l1", "l2"),
  tclose_metric_numeric = c("ks", "wass", "cvm", "ad"),
  na.rm = FALSE
)
}
\arguments{
\item{eval_data}{An \code{eval_data} object or a tibble / data.frame corresponding
to the confidential data.}

\item{qid_keys}{A character vector of quasi-identifying keys. Must be provided
as \code{factor} type variables. Defaults to all factor variables in \code{eval_data}.}

\item{sens_keys}{An optional character vector of sensitive variables of interest.
Must be disjoint from \code{qid_keys}, or \code{FALSE} to provide no l-diversity or t-closeness
metrics. Defaults to the complement of \code{qid_keys}.}

\item{tclose_metric_cat}{A string describing the t-closeness distance metric
between proportions of categorical variables. One of \code{"l1"} (L1 distance),
\code{"l2"} (L2 distance), or \code{"linf"} (L-infinity distance), defaults to \code{"linf"}.}

\item{tclose_metric_numeric}{A string describing the t-closeness distance metric
between numeric variable empirical CDFs. One of \code{"ks"} (Kolmogorov-Smirnov),
\code{"wass"} (Wasserstein L1 distance), \code{"cvm"} (Cramer-von-Mises distance), or
\code{"ad"} (Anderson-Darling), defaults, to \code{"ks"}.}

\item{na.rm}{Boolean if TRUE, will remove \code{NA} values from numeric \code{sens_keys}.}
}
\value{
A \code{baseline_metrics} object.
}
\description{
Compute baseline disclosure risk metrics using confidential data.
}
