% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminator_auc.R
\name{discriminator_auc}
\alias{discriminator_auc}
\title{Compute observed pmse for a single synthetic data set}
\usage{
discriminator_auc(
  postsynth,
  data,
  formula = id ~ .,
  cp = 0.01,
  model = "decision tree"
)
}
\arguments{
\item{postsynth}{A data frame with the synthetic data}

\item{data}{A postsynth object or data frame with confidential data}

\item{formula}{A formula for the predictive model}

\item{cp}{A hyperparameter for rpart}

\item{model}{A character vector with the model name of choice}
}
\value{
joint distributional utility metric (num)
}
\description{
Compute observed pmse for a single synthetic data set
}
