% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_dd_helper.R
\name{.aggregate_qid}
\alias{.aggregate_qid}
\title{Aggregate and count factor-level quasi-identifiers.}
\usage{
.aggregate_qid(df, keys)
}
\arguments{
\item{df}{A data.frame.}

\item{keys}{A character vector of column names.}
}
\value{
A data.frame of aggregated quasi-identifiers.
}
\description{
Aggregate and count factor-level quasi-identifiers.
}
