% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_bivariate.r
\name{.calc_rmi_tibble}
\alias{.calc_rmi_tibble}
\title{Calculate relative mutual information}
\usage{
.calc_rmi_tibble(df, group_by_q)
}
\arguments{
\item{df}{a dataframe containing factor variables}

\item{group_by_q}{optional quoted character string of a variable name to
group the data by. If provided, the correlation matrix will be calculated
for each group separately.}
}
\value{
A long tibble with relative mutual information
}
\description{
Calculate relative mutual information
}
