% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_nse_helper.R
\name{.create_combined_data_pointwise}
\alias{.create_combined_data_pointwise}
\title{Create combined data for pointwise utility statistic evaluation}
\usage{
.create_combined_data_pointwise(
  synth_data,
  conf_data,
  keep_numeric = TRUE,
  group_by_q = NULL,
  weight_var_q = NULL,
  common_vars = FALSE,
  synth_varnames = NULL
)
}
\arguments{
\item{synth_data}{A synthetic data.frame}

\item{conf_data}{A confidential data.frame}

\item{keep_numeric}{A boolean, if TRUE keeps only numeric variables, else keeps
factors and characters. Defaults to TRUE.}

\item{group_by_q}{The quoted name(s) of a (or multiple) grouping variable(s)}

\item{weight_var_q}{A quoted name of a weight variable}

\item{common_vars}{A logical for if only common variables should be kept}

\item{synth_varnames}{A list of variables synthesized to filter on, else \code{NULL}}
}
\value{
A tibble
}
\description{
Create combined data for pointwise utility statistic evaluation
}
