% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_helper.R
\name{.get_target_predictions}
\alias{.get_target_predictions}
\title{Retrieve a predictions tibble from an attribute_target object}
\usage{
.get_target_predictions(target, source, on)
}
\arguments{
\item{target}{An \code{attribute_target} object.}

\item{source}{Either \code{"synthetic"} or \code{"holdout"} - which fitted branch's
predictions to retrieve.}

\item{on}{Either \code{"confidential"}, \code{"train"}, or \code{"test"} - which
predictions tibble to retrieve from the branch.}
}
\value{
A predictions tibble.
}
\description{
Retrieve a predictions tibble from an attribute_target object
}
