% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_na_helper.R
\name{.prep_combined_data_for_na.rm_q}
\alias{.prep_combined_data_for_na.rm_q}
\title{Check whether na.rm is compatible with univariate utilty metrics}
\usage{
.prep_combined_data_for_na.rm_q(
  combined_data,
  na.rm = FALSE,
  drop_zeros = FALSE,
  drop_zeros_exclude = NULL
)
}
\arguments{
\item{combined_data}{A data frame or tibble}

\item{na.rm}{A boolean for whether to ignore missing values}

\item{drop_zeros}{A boolean for whether to ignore zero values in utility metrics}

\item{drop_zeros_exclude}{An optional set of quoted columns on which to drop zeros}
}
\value{
A data frame or tibble with missing and/or zero values set to NA
}
\description{
Check whether na.rm is compatible with univariate utilty metrics
}
