% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_bivariate.r
\name{.relative_mutual_information}
\alias{.relative_mutual_information}
\title{Calculate relative mutual information}
\usage{
.relative_mutual_information(x, y)
}
\arguments{
\item{x}{A numeric or factor vector}

\item{y}{A numeric or factor vector}
}
\value{
A numeric relative mutual information between 0 and 1
}
\description{
Calculate relative mutual information
}
