% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_bivariate.r
\name{.rmi_entropy}
\alias{.rmi_entropy}
\title{Calculate entropy for RMI matrix}
\usage{
.rmi_entropy(p)
}
\arguments{
\item{p}{A vector of numeric probabilities}
}
\value{
a numeric value for entropy of probabilities
}
\description{
Calculate entropy for RMI matrix
}
