% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_dd_helper.R
\name{.target_distribution}
\alias{.target_distribution}
\title{Compute the conditional distribution of a single target variable within
quasi-identifier equivalence classes.}
\usage{
.target_distribution(target, df, qid_keys)
}
\arguments{
\item{target}{A single target column name.}

\item{df}{A data.frame.}

\item{qid_keys}{A character vector of quasi-identifying column names.}
}
\value{
A tibble with columns for \code{qid_keys}, \code{target_var}, \code{target_level},
\code{n}, and \code{prob}.
}
\description{
Compute the conditional distribution of a single target variable within
quasi-identifier equivalence classes.
}
