% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_bivariate.r
\name{.util_bivariate}
\alias{.util_bivariate}
\title{Calculate bivariate statistics  of a confidential data set.}
\usage{
.util_bivariate(
  synth_data = synth_data,
  conf_data,
  statistic,
  use = "everything",
  group_by_q = NULL,
  method = "pearson"
)
}
\arguments{
\item{synth_data}{A data.frame with synthetic data}

\item{conf_data}{A data.frame with the confidential data}

\item{statistic}{a character string specifying which bivariate statistic
to be returned by the function. One of "correlation", "covariance", or "RMI"}

\item{use}{optional character string giving a method for computing
covariances in the presence of missing values. This must be (an abbreviation
of) one of the strings "everything", "all.obs", "complete.obs",
"na.or.complete", or "pairwise.complete.obs".}

\item{group_by_q}{optional quoted character string of a variable name to
group the data by. If provided, the covariance matrix will be calculated
for each group separately.}

\item{method}{optional character string indicating which covariance
is to be computed. One of "pearson" (default), "kendall", or "spearman".}
}
\value{
A long dataset with var1, var2, group_by_q if provided, and covariance
}
\description{
Calculate bivariate statistics  of a confidential data set.
}
