% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_co_ocurrence.R
\name{.util_co_occurrence}
\alias{.util_co_occurrence}
\title{Compare the co-occurrence fit metric of a confidential and synthetic dataset}
\usage{
.util_co_occurrence(synth_data, conf_data, na.rm = FALSE)
}
\arguments{
\item{synth_data}{A data.frame with synthetic data}

\item{conf_data}{A data.frame with the confidential data}

\item{na.rm}{a logical indicating whether missing values should be removed.
Note: values are jointly removed for each pair of variables even if only one
value is missing.}
}
\value{
A \code{list} of fit metrics:
\itemize{
\item \code{co_occurrence_original}: co-occurrence matrix of the original data.
\item \code{co_occurrence_synthetic}: co-occurrence matrix of the synthetic data.
\item \code{co_occurrence_difference}: difference between \code{co_occurrence_synthetic} and
\code{co_occurrence_original}.
\code{co_occurrence_synthetic} and \code{co_occurrence_original}, divided by the number of
cells in the co-occurrence matrix.
\item \code{co_occurrence_difference_mae}: Mean absolute error between
\code{co_occurrence_original} and \code{co_occurrence_synthetic}
\item \code{co_occurrence_difference_rmse}: Root mean squared error between
\code{co_occurrence_original} and \code{co_occurrence_synthetic}
}
}
\description{
Compare the co-occurrence fit metric of a confidential and synthetic dataset
}
