% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ks_distance.R
\name{.util_ks_distance}
\alias{.util_ks_distance}
\title{Calculate the Kolmogorov-Smirnov distance (D) for each numeric variable in
the synthetic and confidential data}
\usage{
.util_ks_distance(synth_data, conf_data, na.rm = FALSE)
}
\arguments{
\item{synth_data}{A data.frame with synthetic data}

\item{conf_data}{A data.frame with the confidential data}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\value{
A tibble with the D and location of the largest distance for each
numeric variable
}
\description{
Calculate the Kolmogorov-Smirnov distance (D) for each numeric variable in
the synthetic and confidential data
}
