% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_moments.R
\name{.util_moments}
\alias{.util_moments}
\title{Calculate summary statistics for original and synthetic data.}
\usage{
.util_moments(
  synth_data,
  conf_data,
  weight_var = 1,
  group_by = NULL,
  drop_zeros = FALSE,
  common_vars = TRUE,
  synth_vars = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{synth_data}{A data.frame with synthetic data}

\item{conf_data}{A data.frame with the confidential data}

\item{weight_var}{An unquoted name of a weight variable}

\item{group_by}{The unquoted name of a (or multiple) grouping variable(s)}

\item{drop_zeros}{A logical for if zeros should be dropped}

\item{common_vars}{A logical for if only common variables should be kept}

\item{synth_vars}{A list of variables synthesized to filter on, else \code{NULL}}

\item{na.rm}{A logical for ignoring \code{NA} values in computations.}
}
\value{
A \code{tibble} of summary statistics.
}
\description{
Calculate summary statistics for original and synthetic data.
}
