% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_percentiles.R
\name{.util_percentiles}
\alias{.util_percentiles}
\title{Calculate summary statistics for original and synthetic data.}
\usage{
.util_percentiles(
  synth_data,
  conf_data,
  probs = c(0.1, 0.5, 0.9),
  weight_var_q = NULL,
  group_by_q = NULL,
  drop_zeros = FALSE,
  common_vars = TRUE,
  synth_varnames = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{synth_data}{A data.frame with synthetic data}

\item{conf_data}{A data.frame with the confidential data}

\item{probs}{A numeric vector of probabilities with values in [0,1]. The
percentiles are interpolated using an empirical CDF. It's possible that the
percentiles are an approximation; especially when weights are used.}

\item{weight_var_q}{A quoted name of a weight variable}

\item{group_by_q}{The quoted name(s) of a (or multiple) grouping variable(s)}

\item{drop_zeros}{A Boolean for if zeros should be dropped}

\item{common_vars}{A logical for if only common variables should be kept.
This option will frequently result in an error because quantile() is strict
about missing values.}

\item{synth_varnames}{A list of variables synthesized to filter on, else \code{NULL}}

\item{na.rm}{A logical for ignoring \code{NA} values in computations.}
}
\value{
A \code{tibble} of summary statistics.
}
\description{
Calculate summary statistics for original and synthetic data.
}
