% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_proportions.R
\name{.util_proportions}
\alias{.util_proportions}
\title{Calculate relative frequency tables for categorical variables}
\usage{
.util_proportions(
  synth_data,
  conf_data,
  weight_var_q = NULL,
  group_by_q = NULL,
  common_vars = TRUE,
  synth_varnames = NULL,
  keep_empty_levels = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{synth_data}{A data.frame with synthetic data}

\item{conf_data}{A data.frame with the confidential data}

\item{weight_var_q}{A quoted name of a weight variable}

\item{group_by_q}{The quoted name(s) of a (or multiple) grouping variable(s)}

\item{common_vars}{A logical for if only common variables should be kept}

\item{synth_varnames}{A list of variables synthesized to filter on, else \code{NULL}}

\item{keep_empty_levels}{A logical for keeping all class levels in the group_by
statements, including missing levels.}

\item{na.rm}{A logical for ignoring \code{NA} values in proportion calculations.}
}
\value{
A tibble with variables, classes, and relative frequencies
}
\description{
Calculate relative frequency tables for categorical variables
}
