% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_totals.R
\name{.util_totals}
\alias{.util_totals}
\title{Calculate totals for original and synthetic data.}
\usage{
.util_totals(
  synth_data,
  conf_data,
  weight_var_q = NULL,
  group_by_q = NULL,
  common_vars = TRUE,
  synth_varnames = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{synth_data}{A data.frame with synthetic data}

\item{conf_data}{A data.frame with the confidential data}

\item{weight_var_q}{A quoted name of a weight variable}

\item{group_by_q}{The quoted name(s) of a (or multiple) grouping variable(s)}

\item{common_vars}{A logical for if only common variables should be kept}

\item{synth_varnames}{A list of variables synthesized to filter on, else \code{NULL}}

\item{na.rm}{A logical for ignoring \code{NA} values in computations.}
}
\value{
A \code{tibble} of totals.
}
\description{
Calculate totals for original and synthetic data.
}
