% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_k_marginal_vars.R
\name{.validate_bins}
\alias{.validate_bins}
\title{Input validation for the number of bins in
.discretize_k_marginal_vars()}
\usage{
.validate_bins(bins)
}
\arguments{
\item{bins}{numeric scalar for number of bins for data}
}
\value{
\code{TRUE} if \code{bins} is a single integer >= 2; otherwise an error is
thrown.
}
\description{
The function returns an error when any of the conditions are met:
\enumerate{
\item bins is not a numeric data type
\item bins is not a scalar
\item bins is an NA value
\item bins is lesser than 2
\item bins is not an integer
}
}
