% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_dd_helper.R
\name{.validate_eval_keys}
\alias{.validate_eval_keys}
\title{Input validation for categorical disclosure risk metrics.}
\usage{
.validate_eval_keys(eval_data, keys = NULL)
}
\arguments{
\item{eval_data}{An \code{eval_data} object.}

\item{keys}{A character vector of column names, or NULL to use all column names.}
}
\description{
Input validation for categorical disclosure risk metrics.
}
