% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_scan.R
\name{is_attribute_scan}
\alias{is_attribute_scan}
\title{Check if object is \code{attribute_scan}}
\usage{
is_attribute_scan(x)
}
\arguments{
\item{x}{object}
}
\value{
A boolean
}
\description{
Check if object is \code{attribute_scan}
}
