% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_target.R
\name{is_attribute_target}
\alias{is_attribute_target}
\title{Check if object is \code{attribute_target}}
\usage{
is_attribute_target(x)
}
\arguments{
\item{x}{object}
}
\value{
A boolean
}
\description{
Check if object is \code{attribute_target}
}
