% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrimination.R
\name{is_discrimination}
\alias{is_discrimination}
\title{Check if object is \code{discrimination}}
\usage{
is_discrimination(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A boolean.
}
\description{
Check if object is \code{discrimination}
}
