% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_data.R
\name{is_eval_data}
\alias{is_eval_data}
\title{Check if obejct is \code{eval_data}}
\usage{
is_eval_data(x)
}
\arguments{
\item{x}{object}
}
\value{
A boolean
}
\description{
Check if obejct is \code{eval_data}
}
