% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmarginals.R
\name{kmarginals}
\alias{kmarginals}
\title{Calculate k-way marginals metric}
\usage{
kmarginals(postsynth, data, k)
}
\arguments{
\item{postsynth}{A postsynth object or tibble with synthetic data}

\item{data}{A data frame with the original data}

\item{k}{The number of marginals to consider}
}
\value{
A numeric MabsDD score
}
\description{
This is the mean absolute difference between cells from a k-way frequency
histogram from the original data and a frequency histogram from the synthetic
histogram. This calculates the mean of many MabsDD.
}
