% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership_inference_test.R
\name{membership_inference_test}
\alias{membership_inference_test}
\title{Run a membership inference test}
\usage{
membership_inference_test(
  postsynth,
  data,
  holdout_data,
  threshold_percentile = NULL
)
}
\arguments{
\item{postsynth}{A postsynth object or tibble with synthetic data}

\item{data}{A data frame with the original data}

\item{holdout_data}{A data frame with holdout data}

\item{threshold_percentile}{Distances below the value associated with this
percentile will be predicted as in the training data}
}
\value{
A list with precision, recall, the confusion matrix, and ROC AUC
}
\description{
Run a membership inference test
}
