% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moments}
\alias{moments}
\title{Calculate summary statistics for original and synthetic data.}
\usage{
moments(postsynth, data, weight_var = NULL, drop_zeros = FALSE)
}
\arguments{
\item{full}{Use all rows (vs only nonzero)? Defaults to \code{TRUE}.}
}
\value{
A \code{data.frame} of summary statistics.
}
\description{
Calculate summary statistics for original and synthetic data.
}
\seealso{
Other utility functions: 
\code{\link{cor_fit}()}
}
\concept{utility functions}
