% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_pmse.R
\name{null_pmse}
\alias{null_pmse}
\title{calculate null pmse}
\usage{
null_pmse(data_orig, times, formula = id ~ ., cp)
}
\arguments{
\item{data_orig}{A data frame with the original data}

\item{times}{The number of permutations}

\item{formula}{A formula for the CART model}

\item{cp}{A hyperparameter for CART}
}
\value{
joint distributional utility metric under the null of two data sets drawn from the same distribution (num)
}
\description{
calculate null pmse
}
