% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguins_conf.R
\docType{data}
\name{penguins_conf}
\alias{penguins_conf}
\title{Penguins data without missing values}
\format{
A data frame with 333 observations and 7 variables.
}
\source{
library(palmerpenguins)
}
\usage{
penguins_conf
}
\description{
Penguins data without missing values
}
\keyword{datasets}
