% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguins_postsynth.R
\docType{data}
\name{penguins_postsynth}
\alias{penguins_postsynth}
\title{A postsynth object for a synthetic penguins data set}
\format{
A postsynth object with 333 observations and 7 variables.
}
\source{
library(palmerpenguins)
}
\usage{
penguins_postsynth
}
\description{
A postsynth object for a synthetic penguins data set
}
\keyword{datasets}
