% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguins_syn_df.R
\docType{data}
\name{penguins_syn_df}
\alias{penguins_syn_df}
\title{A df for a synthetic penguins data set}
\format{
A df with 333 observations and 7 variables.
}
\source{
library(palmerpenguins)
}
\usage{
penguins_syn_df
}
\description{
A df for a synthetic penguins data set
}
\keyword{datasets}
