% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{permutation_distribution}
\alias{permutation_distribution}
\title{Permutation test function}
\usage{
permutation_distribution(df_combined, utility_fun, n_permutations = 500, ...)
}
\description{
Permutation test function
}
\keyword{internal}
