% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plots.R
\name{plot_categorical_bar}
\alias{plot_categorical_bar}
\title{Create bar charts for a categorical random variable.}
\usage{
plot_categorical_bar(joint_data, var_name, cat1_name = NULL, cat2_name = NULL)
}
\arguments{
\item{joint_data}{A data.frame combining rows from confidential and synthetic
data, with the column 'source' identifying the two.}

\item{var_name}{Categorical variable name to plot.}

\item{cat1_name}{Optional categorical variable to group by for subplots.}

\item{cat2_name}{Optional categorical variable to group by for subplots.}
}
\value{
A \code{ggplot2} plot
}
\description{
Create bar charts for a categorical random variable.
}
