% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plots.R
\name{plot_na_matrix}
\alias{plot_na_matrix}
\title{Create side-by-side black/white missingness matrix plots.}
\usage{
plot_na_matrix(eval_data, na_values = NULL)
}
\arguments{
\item{eval_data}{An \code{eval_data} object.}

\item{na_values}{An optional scalar or vector of values (in addition to \code{NA})
that should be treated as missing}
}
\value{
A \code{ggplot2} plot
}
\description{
Create side-by-side black/white missingness matrix plots.
}
