% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmse.R
\name{pmse}
\alias{pmse}
\title{Compute observed pmse for a single synthetic data set}
\usage{
pmse(data_orig, data_syn, formula = id ~ ., cp = 0.01)
}
\arguments{
\item{data_orig}{A data frame with the original data}

\item{data_syn}{A data frame with the synthetic data}

\item{formula}{A formula for the predictive model}

\item{cp}{A hyperparameter for CART}
}
\value{
joint distributional utility metric (num)
}
\description{
Compute observed pmse for a single synthetic data set
}
