% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmse_ratio.R
\name{pmse_ratio}
\alias{pmse_ratio}
\title{Calculate the pMSE using a decision tree}
\usage{
pmse_ratio(
  postsynth,
  data,
  formula = id ~ .,
  times = 100,
  cp = 0.01,
  null_pmse = NULL
)
}
\arguments{
\item{postsynth}{An object of class \code{postsynth}.}

\item{data}{The original data set.}

\item{formula}{A formula for the model for calculating propensity scores.}

\item{times}{The number of bootstrap samples to use when simulating the
null_pmse.}

\item{cp}{The complexity parameter for the decision tree model.}

\item{null_pmse}{An optional null_pmse to save computation time from
simulating the null_pmse.}
}
\value{
A list with the pMSE, null pMSE, and pMSE ratio.
}
\description{
Calculate the pMSE using a decision tree
}
