% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_pred.R
\name{pred_auc}
\alias{pred_auc}
\title{Area under the ROC curve for a \code{pred} object}
\usage{
pred_auc(pred)
}
\arguments{
\item{pred}{A \code{pred} object created by \code{util_pred()}.}
}
\value{
A tibble with one row per \code{model}/\code{data} combination and the ROC
AUC for those predictions.
}
\description{
Area under the ROC curve for a \code{pred} object
}
\seealso{
Other Predictive utility metrics: 
\code{\link{pred_mae}()},
\code{\link{pred_precision}()},
\code{\link{pred_recall}()},
\code{\link{pred_rmse}()},
\code{\link{pred_visualize}()},
\code{\link{util_pred}()}
}
\concept{Predictive utility metrics}
