% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_pred.R
\name{pred_precision}
\alias{pred_precision}
\title{Precision for a \code{pred} object}
\usage{
pred_precision(pred)
}
\arguments{
\item{pred}{A \code{pred} object created by \code{util_pred()}.}
}
\value{
A tibble with one row per \code{model}/\code{data} combination and the
precision for those predictions.
}
\description{
Precision for a \code{pred} object
}
\seealso{
Other Predictive utility metrics: 
\code{\link{pred_auc}()},
\code{\link{pred_mae}()},
\code{\link{pred_recall}()},
\code{\link{pred_rmse}()},
\code{\link{pred_visualize}()},
\code{\link{util_pred}()}
}
\concept{Predictive utility metrics}
