% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_dd_helper.R
\name{prep_discrete_eval_data}
\alias{prep_discrete_eval_data}
\title{Discretize numeric columns for disclosure risk metrics}
\usage{
prep_discrete_eval_data(eval_data, col_map)
}
\arguments{
\item{eval_data}{A data.frame}

\item{col_map}{A list mapping each column to its discretization parameters, one
of either "k" (for specifying the total number of categories) or "width" (for
specifying the width of each bin)}
}
\description{
Discretize numeric columns for disclosure risk metrics
}
