% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrimination.R
\name{print.discrimination}
\alias{print.discrimination}
\title{Print a discrimination object}
\usage{
\method{print}{discrimination}(x, ...)
}
\arguments{
\item{x}{A \code{discrimination} object.}

\item{...}{Additional arguments passed to or from other methods. Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a discrimination object
}
