% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_ci_overlap.R
\name{regression_ci_overlap}
\alias{regression_ci_overlap}
\title{Regression confidence interval overlap}
\usage{
regression_ci_overlap(postsynth, data, formula)
}
\arguments{
\item{postsynth}{A postsynth object or tibble with synthetic data}

\item{data}{A data frame with the original data}

\item{formula}{A formula for the regression model}
}
\value{
A list with the regression confidence interval overlap and estimated
coefficients
}
\description{
Regression confidence interval overlap
}
