% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_constraints.R
\name{util_constraints}
\alias{util_constraints}
\title{Evaluate constraints}
\usage{
util_constraints(
  eval_data,
  constraints_df_num = NULL,
  constraints_df_cat = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{eval_data}{An \code{eval_data} object.}

\item{constraints_df_num}{A numeric constraints data frame in the style of \code{tidysynthesis}.}

\item{constraints_df_cat}{A categorical constraints data frame in the style of \code{tidysynthesis}.}

\item{na.rm}{a logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped
before the computation proceeds.}
}
\value{
A list with data frame with summary statistics for how often constraints apply and how often
constraints are satisfied.
}
\description{
Evaluate constraints
}
