% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_d.R
\name{util_d}
\alias{util_d}
\title{Calculate the KS distance (D) for numeric variables}
\usage{
util_d(postsynth, data)
}
\arguments{
\item{postsynth}{A postsynth object or tibble with synthetic data}

\item{data}{A data frame with the original data}
}
\description{
Calculate the KS distance (D) for numeric variables
}
