% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tails.R
\name{tails}
\alias{tails}
\title{Explore the tails of numeric variables}
\usage{
tails(postsynth, data, n = 10, weight_var = 1, end = "max")
}
\arguments{
\item{postsynth}{A postsynth object or tibble with synthetic data}

\item{data}{A data frame with the original data}

\item{n}{The number of observations to consider for each variable}

\item{weight_var}{An unquoted name of a weight variable}

\item{end}{"min" for minimum values and "max" for maximum values}
}
\value{
A \code{tibble} of summary statistics.
}
\description{
Explore the tails of numeric variables
}
\seealso{
Other utility functions: 
\code{\link{cor_fit}()},
\code{\link{moments}()},
\code{\link{percentiles}()}
}
\concept{utility functions}
