% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_pmse_logistic.R
\name{utility_pmse_logistic}
\alias{utility_pmse_logistic}
\title{pMSE metric using logistic regression}
\usage{
utility_pmse_logistic(df_syn, df_orig)
}
\arguments{
\item{df_syn}{synthesized data frame}

\item{df_orig}{original data frame}
}
\value{
pmse value
}
\description{
pMSE metric using logistic regression
}
\examples{
 # bad
 utility_pmse_logistic(birthwt_ind, birthwt)

 # good
 utility_pmse_logistic(birthwt_dep, birthwt)

}
