% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_pmse_rf.R
\name{utility_pmse_rf}
\alias{utility_pmse_rf}
\title{pMSE metric using random forest}
\usage{
utility_pmse_rf(df_syn, df_orig, test = FALSE, ...)
}
\arguments{
\item{df_syn}{synthesized data frame}

\item{df_orig}{original data frame}
}
\value{
pmse value
}
\description{
pMSE metric using random forest
}
\examples{
 # bad
 utility_pmse_rf(birthwt_ind, birthwt)

 # good
 utility_pmse_rf(birthwt_dep, birthwt)

}
