% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_discriminator_auc.R
\name{add_discriminator_auc}
\alias{add_discriminator_auc}
\title{Add discriminator AUC to discrimination object}
\usage{
add_discriminator_auc(discrimination)
}
\arguments{
\item{discrimination}{A discrimination with added propensities}
}
\value{
A discrimination with discriminator AUC
}
\description{
Add discriminator AUC to discrimination object
}
