% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pmse.R
\name{add_pmse}
\alias{add_pmse}
\title{Add pMSE to discrimination object}
\usage{
add_pmse(discrimination, split = TRUE, group_by_q = NULL)
}
\arguments{
\item{discrimination}{A discrimination object with propensities (likely
added using add_propensities())}

\item{split}{A logical for if the metric should be calculated separately for
the training/testing split. Defaults to TRUE.}

\item{group_by_q}{An optional grouping variable to calculate the pMSE within each group.}
}
\value{
A discrimination object with propensities (likely added using
add_propensities()) with a pMSE
}
\description{
Add pMSE to discrimination object
}
\seealso{
Other Utility metrics: 
\code{\link{add_pmse_ratio}()},
\code{\link{add_propensities}()},
\code{\link{add_propensities_tuned}()},
\code{\link{add_specks}()},
\code{\link{compute_ci_overlap}()},
\code{\link{discrimination}()},
\code{\link{util_ci_overlap}()},
\code{\link{util_co_occurrence}()},
\code{\link{util_ks_distance}()},
\code{\link{util_proportions}()}
}
\concept{Utility metrics}
