% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pmse.R
\name{add_pmse}
\alias{add_pmse}
\title{Add pMSE to discrimination object}
\usage{
add_pmse(discrimination)
}
\arguments{
\item{discrimination}{A discrimination with added propensities}
}
\value{
A discrimination with pMSE
}
\description{
Add pMSE to discrimination object
}
