% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pmse.R
\name{add_pmse}
\alias{add_pmse}
\title{Add pMSE to discrimination object}
\usage{
add_pmse(discrimination, split = TRUE)
}
\arguments{
\item{discrimination}{A discrimination object with propensities (likely
added using add_propensities())}

\item{split}{A logical for if the metric should be calculated separately for
the training/testing split. Defaults to TRUE.}
}
\value{
A discrimination object with propensities (likely added using
add_propensities()) with a pMSE
}
\description{
Add pMSE to discrimination object
}
