% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pmse.R
\name{add_pmse}
\alias{add_pmse}
\title{Add pMSE to discrimination object, with option to assess separately on groups
indicated by a grouping variable}
\usage{
add_pmse(discrimination, group = c(), split = TRUE)
}
\arguments{
\item{discrimination}{A discrimination object with propensities (likely
added using add_propensities())}

\item{group}{A set of variables to group the pmse by}

\item{split}{A logical for if the metric should be calculated separately for
the training/testing split. Defaults to TRUE.}
}
\value{
A discrimination object with propensities (likely added using
add_propensities()) with a pMSE
}
\description{
Add pMSE to discrimination object, with option to assess separately on groups
indicated by a grouping variable
}
\seealso{
Other Utility metrics: 
\code{\link{add_pmse_ratio}()},
\code{\link{add_propensities_tuned}()},
\code{\link{add_propensities}()},
\code{\link{add_specks}()},
\code{\link{discrimination}()},
\code{\link{util_ci_overlap}()},
\code{\link{util_ks_distance}()},
\code{\link{util_proportions}()}
}
\concept{Utility metrics}
