% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pmse_ratio.R
\name{add_pmse_ratio}
\alias{add_pmse_ratio}
\title{Add pMSE ratio to discrimination object}
\usage{
add_pmse_ratio(discrimination, split = TRUE, prop = 3/4, group = c(), times)
}
\arguments{
\item{discrimination}{A discrimination with added propensities}

\item{split}{A logical for if the metric should be calculated separately for
the training/testing split. Defaults to TRUE.}

\item{prop}{The proportion of data to be retained for modeling/analysis in
the training/testing split. The sampling is stratified by the original and
synthetic data.}

\item{group}{A set of variables to group the pmse and null pmse by}

\item{times}{The number of bootstrap samples.}
}
\value{
A discrimination with pMSE
}
\description{
Add pMSE ratio to discrimination object
}
\seealso{
Other Utility metrics: 
\code{\link{add_pmse}()},
\code{\link{add_propensities_tuned}()},
\code{\link{add_propensities}()},
\code{\link{add_specks}()},
\code{\link{discrimination}()},
\code{\link{util_ci_overlap}()},
\code{\link{util_ks_distance}()},
\code{\link{util_proportions}()}
}
\concept{Utility metrics}
