% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_propensities.R
\name{add_propensities}
\alias{add_propensities}
\title{Title}
\usage{
add_propensities(
  discrimination,
  recipe = NULL,
  formula = NULL,
  spec,
  save_fit = TRUE
)
}
\arguments{
\item{save_fit}{is required for the pMSE ratio}

\item{object}{A parsnip model specification or a
\href{http://127.0.0.1:30225/help/library/workflows/help/workflow}{workflows::workflow()}.
No tuning parameters are allowed.}
}
\description{
Title
}
