% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_propensities.R
\name{add_propensities}
\alias{add_propensities}
\title{Add propensities for if an observation belongs to the synthetic data}
\usage{
add_propensities(
  discrimination,
  recipe = NULL,
  formula = NULL,
  spec,
  save_fit = TRUE
)
}
\arguments{
\item{discrimination}{A discrimination object created by discrimination()}

\item{recipe}{A recipe object from library(recipes)}

\item{formula}{A formula for the discriminator model}

\item{spec}{A model object from library(parsnip)}

\item{save_fit}{A logical for if the final model should be saved}
}
\value{
A discrimination object with propensities and a fitted model for
generating propensities
}
\description{
Add propensities for if an observation belongs to the synthetic data
}
\seealso{
Other Utility metrics: 
\code{\link{add_pmse_ratio}()},
\code{\link{add_pmse}()},
\code{\link{add_propensities_tuned}()},
\code{\link{add_specks}()},
\code{\link{discrimination}()},
\code{\link{util_ci_overlap}()},
\code{\link{util_ks_distance}()},
\code{\link{util_proportions}()}
}
\concept{Utility metrics}
