% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_propensities_tuned.R
\name{add_propensities_tuned}
\alias{add_propensities_tuned}
\title{Add propensities for if an observation belongs to the synthetic data}
\usage{
add_propensities_tuned(
  discrimination,
  recipe = NULL,
  formula = NULL,
  spec,
  grid,
  prop = 3/4,
  v = 10,
  save_fit = TRUE
)
}
\arguments{
\item{discrimination}{A discrimination object created by discrimination()}

\item{recipe}{A recipe object from library(recipes). A recipe can't be used
if a formula is provided.}

\item{formula}{A formula for the discriminator model. A formula can't be used
if a recipe is provided.}

\item{spec}{A model object from library(parsnip). If no recipe or formula is
specified, then all variables are used as predictors.}

\item{grid}{a tibble with hyperparameters for tuning}

\item{prop}{The proportion of data to be retained for modeling/analysis in
the training/testing split. The sampling is stratified by the original and
synthetic data.}

\item{v}{The number of partitions of the data set for cross validation.}

\item{save_fit}{A logical for if the final model should be saved}
}
\value{
A discrimination object with propensities and a fitted model for
generating propensities
}
\description{
Add propensities for if an observation belongs to the synthetic data
}
\seealso{
Other Utility metrics:
\code{\link[=add_pmse]{add_pmse()}},
\code{\link[=add_pmse_ratio]{add_pmse_ratio()}},
\code{\link[=add_propensities]{add_propensities()}},
\code{\link[=add_specks]{add_specks()}},
\code{\link[=discrimination]{discrimination()}},
\code{\link[=util_ci_overlap]{util_ci_overlap()}},
\code{\link[=util_co_occurrence]{util_co_occurrence()}},
\code{\link[=util_ks_distance]{util_ks_distance()}},
\code{\link[=util_proportions]{util_proportions()}}
}
\concept{Utility metrics}
