% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_propensities_tuned.R
\name{add_propensities_tuned}
\alias{add_propensities_tuned}
\title{Add propensities for if an observation belongs to the synthetic data}
\usage{
add_propensities_tuned(
  discrimination,
  recipe = NULL,
  formula = NULL,
  spec,
  grid,
  save_fit = TRUE
)
}
\arguments{
\item{grid}{a tibble with hyperparameters for tuning}

\item{save_fit}{is required for the pMSE ratio}

\item{object}{A parsnip model specification or a
\href{https://workflows.tidymodels.org/}{workflows::workflow()}.
No tuning parameters are allowed.}
}
\value{
A discrimination object with propensities and a fitted model for
generating propensities
}
\description{
Add propensities for if an observation belongs to the synthetic data
}
