% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_specks.R
\name{add_specks}
\alias{add_specks}
\title{Add SPECKS to discrimination object}
\usage{
add_specks(discrimination, group = c(), split = TRUE)
}
\arguments{
\item{discrimination}{A discrimination with added propensities}

\item{group}{A list of variable names to group by}

\item{split}{A logical for if the metric should be calculated separately for
the training/testing split. Defaults to TRUE.}
}
\value{
A discrimination with SPECKS
}
\description{
Add SPECKS to discrimination object
}
\seealso{
Other Utility metrics:
\code{\link[=add_pmse]{add_pmse()}},
\code{\link[=add_pmse_ratio]{add_pmse_ratio()}},
\code{\link[=add_propensities]{add_propensities()}},
\code{\link[=add_propensities_tuned]{add_propensities_tuned()}},
\code{\link[=discrimination]{discrimination()}},
\code{\link[=util_ci_overlap]{util_ci_overlap()}},
\code{\link[=util_co_occurrence]{util_co_occurrence()}},
\code{\link[=util_ks_distance]{util_ks_distance()}},
\code{\link[=util_proportions]{util_proportions()}}
}
\concept{Utility metrics}
