% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_specks.R
\name{add_specks}
\alias{add_specks}
\title{Add SPECKS to discrimination object}
\usage{
add_specks(discrimination, split = TRUE)
}
\arguments{
\item{discrimination}{A discrimination with added propensities}

\item{split}{A logical for if the metric should be calculated separately for
the training/testing split. Defaults to TRUE.}
}
\value{
A discrimination with SPECKS
}
\description{
Add SPECKS to discrimination object
}
