% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_scan.R
\name{attribute_scan}
\alias{attribute_scan}
\title{Create an attribute_scan object for non-parametric discovery attack metrics}
\usage{
attribute_scan(eval_data, qid_keys, target_keys = NULL)
}
\arguments{
\item{eval_data}{An \code{eval_data} object.}

\item{qid_keys}{A character vector of quasi-identifying keys. Must be \code{factor}
type variables present in both the confidential and synthetic data.}

\item{target_keys}{An optional character vector of target variables of interest.
Must be disjoint from \code{qid_keys} and \code{factor} type. Defaults to the complement
of \code{qid_keys} among the factor columns of \code{eval_data$conf_data}.}
}
\value{
An \code{attribute_scan} object. If \code{eval_data$holdout_data} is supplied,
the object also includes a \code{holdout} element with the same structure as
\code{confidential} and \code{synthetic}.
}
\description{
Create an attribute_scan object for non-parametric discovery attack metrics
}
