% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plots.R
\name{create_cormat_plot}
\alias{create_cormat_plot}
\title{Create a correlation heatmap for numeric random variables.}
\usage{
create_cormat_plot(data, cor_method = "pearson", group_by_q = NULL)
}
\arguments{
\item{data}{A data.frame/}

\item{cor_method}{A correlation method to pass to \verb{stats::cor(., method=<cor_method>)}}
}
\value{
A \code{ggplot2} plot
}
\description{
Create a correlation heatmap for numeric random variables.
}
