% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_mit.R
\name{disc_mit}
\alias{disc_mit}
\title{Run a membership inference test}
\usage{
disc_mit(postsynth, data, holdout_data, threshold_percentile = NULL)
}
\arguments{
\item{postsynth}{A postsynth object or tibble with synthetic data generated from the data input}

\item{data}{A data frame with a subset of the original data}

\item{holdout_data}{A dataframe with observations similar to the original but
not used to train the synthesizer. The data should have the same variables as
postsynth.}

\item{threshold_percentile}{Distances below the value associated with this
percentile will be predicted as in the training data. If the
threshold_percentile is not provided, the function calculates it with the
following formula: \code{nrow(data)/(nrow(data) + nrow(holdout_data))}}
}
\value{
A list with precision, recall, the confusion matrix, and ROC AUC
}
\description{
Run a membership inference test
}
