% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrimination.R
\name{discrimination}
\alias{discrimination}
\title{Combine synthetic data and data for a discriminant based metric}
\usage{
discrimination(eval_data)
}
\arguments{
\item{eval_data}{An \code{eval_data} object.}
}
\value{
A list of class discrimination
}
\description{
Combine synthetic data and data for a discriminant based metric
}
\seealso{
Other Utility metrics: 
\code{\link{add_pmse}()},
\code{\link{add_pmse_ratio}()},
\code{\link{add_propensities}()},
\code{\link{add_propensities_tuned}()},
\code{\link{add_specks}()},
\code{\link{compute_ci_overlap}()},
\code{\link{util_ci_overlap}()},
\code{\link{util_co_occurrence}()},
\code{\link{util_ks_distance}()},
\code{\link{util_proportions}()}
}
\concept{Utility metrics}
