% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrimination.R
\name{discrimination}
\alias{discrimination}
\title{Combine synthetic data and data for a discriminant based metric}
\usage{
discrimination(postsynth, data)
}
\arguments{
\item{postsynth}{A postsynth object from tidysynthesis or a tibble}

\item{data}{an original (observed) data set.}
}
\description{
Combine synthetic data and data for a discriminant based metric
}
