% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_corr_fit.R
\name{.lower_triangle}
\alias{.lower_triangle}
\title{Calculate the lower triangle of a correlation matrix for numeric random variables.}
\usage{
.lower_triangle(x, use, group_by_q = NULL, method = "pearson")
}
\arguments{
\item{x}{A data.frame}

\item{use}{optional character string giving a method for computing
covariances in the presence of missing values. This must be (an abbreviation
of) one of the strings "everything", "all.obs", "complete.obs",
"na.or.complete", or "pairwise.complete.obs".}

\item{group_by_q}{optional quoted character string of a variable name to
group the data by. If provided, the correlation matrix will be calculated
for each group separately.}

\item{method}{optional character string indicating which correlation
coefficient is to be computed. One of "pearson" (default), "kendall", or
"spearman".}
}
\value{
A data.frame with columns for the variable pairs and their correlation values.
}
\description{
Calculate the lower triangle of a correlation matrix for numeric random variables.
}
