% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_bivariate.r
\name{.lower_triangle}
\alias{.lower_triangle}
\title{Calculate the lower triangle of a correlation or covariance matrix for
numeric random variables.}
\usage{
.lower_triangle(x, statistic, use, group_by_q = NULL, method = "pearson")
}
\arguments{
\item{x}{A data.frame}

\item{statistic}{a character string specifying which bivariate statistic
to be returned by the function. One of "correlation" or "covariance"}

\item{use}{optional character string giving a method for computing
covariances in the presence of missing values. This must be (an abbreviation
of) one of the strings "everything", "all.obs", "complete.obs",
"na.or.complete", or "pairwise.complete.obs".}

\item{group_by_q}{optional quoted character string of a variable name to
group the data by. If provided, the correlation/covariance matrix will be calculated
for each group separately.}

\item{method}{optional character string indicating which correlation/covariance
coefficient is to be computed. One of "pearson" (default),
"kendall", or "spearman".}
}
\value{
A data.frame with columns for deduplicated variable pairs and their
correlation or covariance values. Diagonal variable pairs (e.g. var1 = a
& var2 = a) are dropped from correlation matrices and preserved in
covariance matrices
}
\description{
Calculate the lower triangle of a correlation or covariance matrix for
numeric random variables.
}
