% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_bivariate.r
\name{.util_bivariate}
\alias{.util_bivariate}
\title{Calculate bivariate statistics of a confidential data set.}
\usage{
.util_bivariate(
  synth_data = synth_data,
  conf_data,
  statistic,
  use = "everything",
  group_by_q = NULL,
  method = "pearson"
)
}
\arguments{
\item{synth_data}{A data.frame with synthetic data}

\item{conf_data}{A data.frame with the confidential data}

\item{statistic}{A character string specifying which bivariate statistic.
One of "correlation", "covariance", or "rmi"}

\item{use}{Optional character string giving a method for computing correlations
or covariances in the presence of missing values. This must be (an abbreviation
of) one of the strings "everything", "all.obs", "complete.obs",
"na.or.complete", or "pairwise.complete.obs". Ignored for statistic = "rmi"}

\item{group_by_q}{Optional quoted character string of a variable name to
group the data by. If provided, the selected statistic will be calculated
for each group separately.}

\item{method}{Optional character string indicating which correlation/covariance
is to be computed. One of "pearson" (default), "kendall", or "spearman".
Ignored for statistic = "rmi"}
}
\value{
A \code{list} of fit metrics:
\itemize{
\item \verb{<statistic>_original}: \if{html}{\out{<statistic>}} values from original data of pairs of
numeric variables that appear in both original and synthetic datasets,
formatted in a long tibble
\item \verb{<statistic>_synthetic}: \if{html}{\out{<statistic>}} values from synthetic data of pairs of
numeric variables that appear in both original and synthetic datasets,
formatted in a long tibble
\item \verb{<statistic>_difference}: difference between \verb{<statistic>_synthetic} and
\verb{<statistic>_original}, formatted in a long tibble.
\item \verb{<statistic>_fit}: square root of the sum of squared differences between
\verb{<statistic>_synthetic} and \verb{<statistic>_original}, divided by the number of
????
\item \verb{<statistic>_difference_mae}: mean absolute error of pairwise differences.
\item \verb{<statistic>_difference_rmse}: root mean squared error of pairwise differences.
}
}
\description{
Calculate bivariate statistics of a confidential data set.
}
