% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_corr_fit.R
\name{.util_corr_fit}
\alias{.util_corr_fit}
\title{Calculate the correlation fit metric of a confidential data set.}
\usage{
.util_corr_fit(synth_data, conf_data, use = "everything", group_by_q = NULL)
}
\arguments{
\item{synth_data}{A data.frame with synthetic data}

\item{conf_data}{A data.frame with the confidential data}

\item{use}{optional character string giving a method for computing
covariances in the presence of missing values. This must be (an abbreviation
of) one of the strings "everything", "all.obs", "complete.obs",
"na.or.complete", or "pairwise.complete.obs".}

\item{group_by_q}{optional quoted character string of a variable name to
group the data by. If provided, the correlation fit metric will be calculated
for each group separately.}
}
\value{
A \code{list} of fit metrics:
\itemize{
\item \code{correlation_original}: correlation matrix of the original data.
\item \code{correlation_synthetic}: correlation matrix of the synthetic data.
\item \code{correlation_difference}: difference between \code{correlation_synthetic} and
\code{correlation_original}.
\item \code{correlation_fit}: square root of the sum of squared differences between
\code{correlation_synthetic} and \code{correlation_original}, divided by the number of
cells in the correlation matrix.
}
}
\description{
Calculate the correlation fit metric of a confidential data set.
}
