% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_k_marginals.R
\name{.util_k_marginals}
\alias{.util_k_marginals}
\title{Worker function for the k-marginals metric}
\usage{
.util_k_marginals(synth_data, conf_data, k)
}
\arguments{
\item{synth_data}{A tibble with synthetic data.}

\item{conf_data}{A tibble with confidential data.}

\item{k}{Scalar order of the k-marginal (valid range = 1:3).}
}
\value{
A value in range \link{0, 1000} where a higher value denotes lower MabsDDs
and consequently greater similarity between confidential and synthetic data.
}
\description{
This worker function takes a specified k-marginal and calculates
all unique k-combinations of variables shared by the input data. For each
combination, tibbles containing unique combinations of observed value levels
are created with the marginal probabilities for each cell for both
synthetic and confidential data included. The mean absolute
difference (MabsDD) between the synthetic and confidential data's marginal
probabilities is computed for each k-combination. The MabsDDs are
averaged into a scalar before being rescaled (1 - mean) * 1000.
}
