% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moments}
\alias{moments}
\title{Calculate summary statistics for original and synthetic data.}
\usage{
moments(postsynth, data, weight_var = NULL, drop_zeros = FALSE)
}
\arguments{
\item{postsynth}{A postsynth object or tibble with synthetic data}

\item{data}{A data frame with the original data}

\item{weight_var}{An unquoted name of a weight variable}

\item{drop_zeros}{A Boolean for if zeros should be dropped}
}
\value{
A \code{tibble} of summary statistics.
}
\description{
Calculate summary statistics for original and synthetic data.
}
\seealso{
Other utility functions: 
\code{\link{cor_fit}()},
\code{\link{tails}()}
}
\concept{utility functions}
