% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_mit.R
\name{nn_membership_inference}
\alias{nn_membership_inference}
\title{Perform a nearest-neighbor membership inference test on one synthetic dataset}
\usage{
nn_membership_inference(
  synth_data,
  conf_data,
  holdout_data,
  threshold_percentile = NULL,
  summary = TRUE
)
}
\arguments{
\item{synth_data}{A dataframe with synthetic data generated from the data input}

\item{conf_data}{A data frame with a subset of the original data}

\item{holdout_data}{A dataframe with observations similar to the original but
not used to train the synthesizer. The data should have the same variables as
postsynth.}

\item{threshold_percentile}{Distances below the value associated with this
percentile will be predicted as in the training data. If the
threshold_percentile is not provided, the function calculates it with the
following formula: \code{nrow(data)/(nrow(data) + nrow(holdout_data))}}

\item{summary}{Boolean if TRUE, returns summary statistics, if FALSE, returns
two disaggregated dataframes of individual distances and ROC curve points.}
}
\value{
Either a list with precision, recall, the confusion matrix, and ROC AUC
or a list with two disaggregated dataframes (if summary = FALSE).
}
\description{
Perform a nearest-neighbor membership inference test on one synthetic dataset
}
