% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plots.R
\name{plot_cormat}
\alias{plot_cormat}
\title{Create side-by-side correlation heatmaps for numeric random variables.}
\usage{
plot_cormat(eval_data, cor_method = "pearson", group_by_q = NULL)
}
\arguments{
\item{eval_data}{An \code{eval_data} object.}

\item{cor_method}{A correlation method to pass to \verb{stats::cor(., method=<cor_method>)}}
}
\value{
A \code{ggplot2} plot
}
\description{
Create side-by-side correlation heatmaps for numeric random variables.
}
