% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plots.R
\name{plot_cormat}
\alias{plot_cormat}
\title{Create side-by-side correlation heatmaps for numeric random variables.}
\usage{
plot_cormat(conf_data, synth_data, cor_method = "pearson")
}
\arguments{
\item{conf_data}{Confidential data}

\item{synth_data}{Synthetic data}

\item{cor_method}{A correlation method to pass to \verb{stats::cor(., method=<cor_method>)}}
}
\value{
A \code{ggplot2} plot
}
\description{
Create side-by-side correlation heatmaps for numeric random variables.
}
