% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plots.R
\name{plot_numeric_hist_kde}
\alias{plot_numeric_hist_kde}
\title{Create a histogram + KDE estimate for a numeric variable.}
\usage{
plot_numeric_hist_kde(joint_data, var_name, cat1_name = NULL, cat2_name = NULL)
}
\arguments{
\item{joint_data}{A data.frame combining rows from confidential and synthetic
data, with the column 'source' identifying the two.}

\item{var_name}{Numeric variable name to plot.}

\item{cat1_name}{Optional categorical variable to group by for subplots.}

\item{cat2_name}{Optional categorical variable to group by for subplots.}
}
\value{
A \code{ggplot2} plot
}
\description{
Create a histogram + KDE estimate for a numeric variable.
}
