% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_pred.R
\name{pred_mae}
\alias{pred_mae}
\title{Mean absolute error for a \code{pred} object}
\usage{
pred_mae(pred)
}
\arguments{
\item{pred}{A \code{pred} object created by \code{util_pred()}.}
}
\value{
A tibble with one row per data source (\code{confidential}/\code{synthetic})
and the MAE of predictions from that source's model.
}
\description{
Mean absolute error for a \code{pred} object
}
\seealso{
Other Predictive utility metrics:
\code{\link[=pred_auc]{pred_auc()}},
\code{\link[=pred_precision]{pred_precision()}},
\code{\link[=pred_recall]{pred_recall()}},
\code{\link[=pred_rmse]{pred_rmse()}},
\code{\link[=pred_visualize]{pred_visualize()}},
\code{\link[=util_pred]{util_pred()}}
}
\concept{Predictive utility metrics}
