% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_pred.R
\name{pred_recall}
\alias{pred_recall}
\title{Recall for a \code{pred} object}
\usage{
pred_recall(pred)
}
\arguments{
\item{pred}{A \code{pred} object created by \code{util_pred()}.}
}
\value{
A tibble with one row per \code{model}/\code{data} combination and the
recall for those predictions.
}
\description{
Recall for a \code{pred} object
}
\seealso{
Other Predictive utility metrics:
\code{\link[=pred_auc]{pred_auc()}},
\code{\link[=pred_mae]{pred_mae()}},
\code{\link[=pred_precision]{pred_precision()}},
\code{\link[=pred_rmse]{pred_rmse()}},
\code{\link[=pred_visualize]{pred_visualize()}},
\code{\link[=util_pred]{util_pred()}}
}
\concept{Predictive utility metrics}
