% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_pred.R
\name{pred_rmse}
\alias{pred_rmse}
\title{Root mean squared error for a \code{pred} object}
\usage{
pred_rmse(pred)
}
\arguments{
\item{pred}{A \code{pred} object created by \code{util_pred()}.}
}
\value{
A tibble with one row per data source (\code{confidential}/\code{synthetic})
and the RMSE of predictions from that source's model.
}
\description{
Root mean squared error for a \code{pred} object
}
\seealso{
Other Predictive utility metrics: 
\code{\link{pred_auc}()},
\code{\link{pred_mae}()},
\code{\link{pred_precision}()},
\code{\link{pred_recall}()},
\code{\link{pred_visualize}()},
\code{\link{util_pred}()}
}
\concept{Predictive utility metrics}
