% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_pred.R
\name{pred_visualize}
\alias{pred_visualize}
\title{Scatterplot comparing confidential and synthetic predictions for a \code{pred} object}
\usage{
pred_visualize(pred, coord_equal = TRUE)
}
\arguments{
\item{pred}{A \code{pred} object created by \code{util_pred()}.}

\item{coord_equal}{Boolean for whether to fix the aspect ratio to 1:1 with
\code{ggplot2::coord_equal()}. Defaults to \code{TRUE}.}
}
\value{
A \code{ggplot2} plot.
}
\description{
For regression models, plots confidential predictions (x-axis) against
synthetic predictions (y-axis). For classification models, plots
confidential predicted probabilities (x-axis) against synthetic predicted
probabilities (y-axis), colored by whether the predicted classes agree.
Only binary classification models are supported.
}
\seealso{
Other Predictive utility metrics: 
\code{\link{pred_auc}()},
\code{\link{pred_mae}()},
\code{\link{pred_precision}()},
\code{\link{pred_recall}()},
\code{\link{pred_rmse}()},
\code{\link{util_pred}()}
}
\concept{Predictive utility metrics}
