% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_equivalence_class_size.R
\name{scan_equivalence_class_size}
\alias{scan_equivalence_class_size}
\title{Summarize the size distribution of equivalence classes}
\usage{
scan_equivalence_class_size(attribute_scan)
}
\arguments{
\item{attribute_scan}{An \code{attribute_scan} object.}
}
\value{
A tibble with columns \code{source} (\code{"confidential"} or \code{"synthetic"}),
\code{n_classes} (the number of observed equivalence classes), \code{min_size},
\code{mean_size}, \code{median_size}, and \code{max_size} (summary statistics of the
observed equivalence class sizes).
}
\description{
Summarize the size distribution of equivalence classes
}
