% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_k_anonymity.R
\name{scan_k_anonymity}
\alias{scan_k_anonymity}
\title{Calculate the minimum equivalence class size defined by the quasi-identifiers}
\usage{
scan_k_anonymity(attribute_scan)
}
\arguments{
\item{attribute_scan}{An \code{attribute_scan} object.}
}
\value{
A tibble with columns \code{source} (\code{"confidential"} or \code{"synthetic"}) and
\code{k} (the minimum observed equivalence class size).
}
\description{
Calculate the minimum equivalence class size defined by the quasi-identifiers
}
