% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_t_closeness.R
\name{scan_t_closeness}
\alias{scan_t_closeness}
\title{Measure how closely each equivalence class target distribution matches the
overall target distribution}
\usage{
scan_t_closeness(attribute_scan, metric = c("linf", "l1", "l2"))
}
\arguments{
\item{attribute_scan}{An \code{attribute_scan} object.}

\item{metric}{A string describing the distance metric between the equivalence
class and overall target distributions. One of \code{"linf"} (L-infinity distance),
\code{"l1"} (L1 distance), or \code{"l2"} (L2 distance), defaults to \code{"linf"}.}
}
\value{
A tibble with columns \code{source} (\code{"confidential"} or \code{"synthetic"}),
\code{target_var}, \code{key_id}, the quasi-identifying key columns, and \code{t_closeness}
(the distance between the equivalence class and overall target distributions).
}
\description{
Measure how closely each equivalence class target distribution matches the
overall target distribution
}
