% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheval-package.R
\docType{package}
\name{syntheval-package}
\alias{syntheval}
\alias{syntheval-package}
\title{syntheval: A set of tools for evaluating synthetic data utility and disclosure risk}
\description{
A set of tools for evaluating the utility and disclosure risks of synthetic data. The package works well with `library(tidysynthesis)` and on its own in other data privacy applications.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/UI-Research/syntheval/issues}
}

}
\author{
\strong{Maintainer}: Aaron R. Williams \email{awilliams@urban.org} (\href{https://orcid.org/0000-0001-5564-1938}{ORCID})

Other contributors:
\itemize{
  \item Gabe Morrison \email{gmorrison@urban.org} [contributor]
  \item Elyse McFalls \email{emcfalls@urban.org} [contributor]
}

}
\keyword{internal}
