% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_bivariate.r
\name{util_bivariate}
\alias{util_bivariate}
\title{Calculate the covariance matrix of a confidential data set.}
\usage{
util_bivariate(
  eval_data,
  statistic,
  use = "everything",
  group_by_q = NULL,
  method = "pearson"
)
}
\arguments{
\item{eval_data}{An \code{eval_data} object}

\item{statistic}{a character string specifying which bivariate statistic
to be returned by the function. One of "correlation", "covariance", or "RMI"}

\item{use}{Optional character string giving a method for computing
covariances in the presence of missing values. This must be (an abbreviation
of) one of the strings "everything", "all.obs", "complete.obs",
"na.or.complete", or "pairwise.complete.obs".}

\item{group_by_q}{Optional quoted character string of a variable name to
group the data by. If provided, the covariance matrix will be calculated
for each group separately.}

\item{method}{optional character string indicating which covariance is to
be computed. One of "pearson" (default), "kendall", or
"spearman".}
}
\value{
A long dataset with var1, var2, group_by_q if provided, and covariance
}
\description{
Calculate the covariance matrix of a confidential data set.
}
\seealso{
Other utility metrics: 
\code{\link{util_corr_fit}()},
\code{\link{util_moments}()},
\code{\link{util_percentiles}()},
\code{\link{util_totals}()}
}
\concept{utility metrics}
