% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ci_overlap.R
\name{util_ci_overlap}
\alias{util_ci_overlap}
\title{Regression confidence interval overlap}
\usage{
util_ci_overlap(eval_data, formula, family = NULL)
}
\arguments{
\item{eval_data}{An \code{eval_data} object}

\item{formula}{A formula for a linear regression model}

\item{family}{An optional family argument, exactly as accepted by
\code{stats::glm()}. When \code{NULL} (the default), \code{stats::lm()} is used to fit
the model.}
}
\value{
A list of two dataframes (one per each synthetic data replicate):
\itemize{
\item \code{ci_overlap}: one row per model parameter with utility metrics.
\itemize{
\item \code{overlap }: symmetric overlap metric, calculated as the average of the
interval overlap contained in the synthetic confidence interval and the
interval overlap contained in the confidential confidence interval.
\item \code{coef_diff}: synthetic parameter estimate - confidential parameter estimate
\item \code{std_coef_diff}: \code{coef_diff} divided by the standard error for the confidential data.
\item \code{sign_match}: boolean if the synthetic and confidential parameter estimates have the same sign.
\item \code{significance_match}: boolean if the null hypothesis test where the
parameter is 0 has p-value less than .05 agrees in both confidential and
synthetic data.
\item \code{ss}: boolean if both \code{sign_match} and \code{significance_match} are true.
\item \code{sso}: boolean if \code{sign_match} is true and \code{overlap} is positive.
}
\item \code{coef_diff}: one row per model parameter and data source (confidential or
synthetic) listing parameter estimates, standard errors, test statistics,
p-values for null hypothesis tests, and 95\% confidence interval bounds.
}
}
\description{
This is the main function for computing confidence interval overlap: it
fits models on the confidential and synthetic data and compares them. If
you already have fitted models to compare, use \code{\link[=compute_ci_overlap]{compute_ci_overlap()}}
instead.
}
\examples{
conf_data <- mtcars
synth_data <- mtcars \%>\% 
   dplyr::slice_sample(n = nrow(mtcars) / 2)
 
eval_data <- eval_data(conf_data, synth_data)

util_ci_overlap(
  eval_data,
  mpg ~ disp + vs + am
)

util_ci_overlap(
  eval_data,
  am ~ disp + vs,
  family = binomial
)

}
\seealso{
Other Utility metrics: 
\code{\link{add_pmse}()},
\code{\link{add_pmse_ratio}()},
\code{\link{add_propensities}()},
\code{\link{add_propensities_tuned}()},
\code{\link{add_specks}()},
\code{\link{compute_ci_overlap}()},
\code{\link{discrimination}()},
\code{\link{util_co_occurrence}()},
\code{\link{util_ks_distance}()},
\code{\link{util_proportions}()}
}
\concept{Utility metrics}
