% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ci_overlap.R
\name{util_ci_overlap}
\alias{util_ci_overlap}
\title{Regression confidence interval overlap}
\usage{
util_ci_overlap(postsynth, data, formula)
}
\arguments{
\item{postsynth}{A postsynth object or tibble with synthetic data}

\item{data}{A data frame with the original data}

\item{formula}{A formula for a linear regression model}
}
\value{
A list with the regression confidence interval overlap and estimated
coefficients
}
\description{
Regression confidence interval overlap
}
\seealso{
Other Utility metrics: 
\code{\link{add_pmse_ratio}()},
\code{\link{add_pmse}()},
\code{\link{add_propensities_tuned}()},
\code{\link{add_propensities}()},
\code{\link{add_specks}()},
\code{\link{discrimination}()},
\code{\link{util_ks_distance}()},
\code{\link{util_proportions}()}
}
\concept{Utility metrics}
