% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_co_ocurrence.R
\name{util_co_occurrence}
\alias{util_co_occurrence}
\title{Compare the co-occurrence fit metric of a confidential and synthetic dataset}
\usage{
util_co_occurrence(eval_data, na.rm = FALSE)
}
\arguments{
\item{eval_data}{An \code{eval_data} object}

\item{na.rm}{a logical indicating whether missing values should be removed.
Note: values are jointly removed for each pair of variables even if only one
value is missing.}
}
\value{
A \code{list} of fit metrics (one per each synthetic data replicate)::
\itemize{
\item \code{co_occurrence_original}: co-occurrence matrix of the original data.
\item \code{co_occurrence_synthetic}: co-occurrence matrix of the synthetic data.
\item \code{co_occurrence_difference}: difference between \code{co_occurrence_synthetic} and
\code{co_occurrence_original}.
\code{co_occurrence_synthetic} and \code{co_occurrence_original}, divided by the number of
cells in the co-occurrence matrix.
\item \code{co_occurrence_difference_mae}: Mean absolute error between
\code{co_occurrence_original} and \code{co_occurrence_synthetic}
\item \code{co_occurrence_difference_rmse}: Root mean squared error between
\code{co_occurrence_original} and \code{co_occurrence_synthetic}
}
}
\description{
Compare the co-occurrence fit metric of a confidential and synthetic dataset
}
\seealso{
Other Utility metrics:
\code{\link[=add_pmse]{add_pmse()}},
\code{\link[=add_pmse_ratio]{add_pmse_ratio()}},
\code{\link[=add_propensities]{add_propensities()}},
\code{\link[=add_propensities_tuned]{add_propensities_tuned()}},
\code{\link[=add_specks]{add_specks()}},
\code{\link[=compute_ci_overlap]{compute_ci_overlap()}},
\code{\link[=discrimination]{discrimination()}},
\code{\link[=util_ci_overlap]{util_ci_overlap()}},
\code{\link[=util_ks_distance]{util_ks_distance()}},
\code{\link[=util_proportions]{util_proportions()}}
}
\concept{Utility metrics}
