% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_co_ocurrence.R
\name{util_co_occurrence}
\alias{util_co_occurrence}
\title{Calculate the co-occurrence fit metric of a confidential data set.}
\usage{
util_co_occurrence(postsynth, data)
}
\arguments{
\item{postsynth}{A postsynth object from tidysynthesis or a tibble}

\item{data}{an original (observed) data set.}
}
\value{
A \code{list} of fit metrics:
\itemize{
\item \code{co_occurrence_original}: co-occurrence matrix of the original data.
\item \code{co_occurrence_synthetic}: co-occurrence matrix of the synthetic data.
\item \code{co_occurrence_difference}: difference between \code{co_occurrence_synthetic} and
\code{co_occurrence_original}.
\code{co_occurrence_synthetic} and \code{co_occurrence_original}, divided by the number of
cells in the co-occurrence matrix.
\item \code{co_occurrence_difference_mae}: Mean absolute error between
\code{co_occurrence_original} and \code{co_occurrence_synthetic}
\item \code{co_occurrence_difference_rmse}: Root mean squared error between
\code{co_occurrence_original} and \code{co_occurrence_synthetic}
}
}
\description{
Calculate the co-occurrence fit metric of a confidential data set.
}
\seealso{
Other utility metrics: 
\code{\link{util_corr_fit}()},
\code{\link{util_moments}()},
\code{\link{util_percentiles}()},
\code{\link{util_tails}()},
\code{\link{util_totals}()}
}
\concept{utility metrics}
