% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_corr_fit.R
\name{util_corr_fit}
\alias{util_corr_fit}
\title{Calculate the correlation fit metric of a confidential data set.}
\usage{
util_corr_fit(eval_data, use = "everything")
}
\arguments{
\item{eval_data}{An \code{eval_data} object}

\item{use}{optional character string giving a method for computing
covariances in the presence of missing values. This must be (an abbreviation
of) one of the strings "everything", "all.obs", "complete.obs",
"na.or.complete", or "pairwise.complete.obs".}
}
\value{
A \code{list} of fit metrics (one per each synthetic data replicate):
\itemize{
\item \code{correlation_original}: correlation matrix of the original data.
\item \code{correlation_synthetic}: correlation matrix of the synthetic data.
\item \code{correlation_difference}: difference between \code{correlation_synthetic} and
\code{correlation_original}.
\item \code{correlation_fit}: square root of the sum of squared differences between
\code{correlation_synthetic} and \code{correlation_original}, divided by the number of
cells in the correlation matrix.
}
}
\description{
Calculate the correlation fit metric of a confidential data set.
}
\seealso{
Other utility metrics:
\code{\link[=util_moments]{util_moments()}},
\code{\link[=util_na_amount]{util_na_amount()}},
\code{\link[=util_na_cluster]{util_na_cluster()}},
\code{\link[=util_percentiles]{util_percentiles()}},
\code{\link[=util_totals]{util_totals()}}
}
\concept{utility metrics}
