% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_corr_fit.R
\name{util_corr_fit}
\alias{util_corr_fit}
\title{Calculate the correlation fit metric of a confidential data set.}
\usage{
util_corr_fit(postsynth, data, group_by = NULL)
}
\arguments{
\item{postsynth}{A postsynth object from tidysynthesis or a tibble}

\item{data}{an original (observed) data set.}

\item{group_by}{The unquoted name of a (or multiple) grouping variable(s)}
}
\value{
A \code{list} of fit metrics:
\itemize{
\item \code{correlation_data}: A \code{tibble} of the correlations among the
numeric variables for the actual and synthetic data
\item \code{correlation_fit}: square root of the sum of squared differences between
the synthetic and original data, divided by the number of
cells in the correlation matrix.
\item \code{correlation_difference_mae}: the mean of the absolute correlation
differences between the actual and synthetic data
\item \code{correlation_difference_rmse}: the root mean of the squared correlation
differences between the actual and synthetic data
}
}
\description{
Calculate the correlation fit metric of a confidential data set.
}
\seealso{
Other utility functions: 
\code{\link{util_co_occurrence}()},
\code{\link{util_moments}()},
\code{\link{util_percentiles}()},
\code{\link{util_tails}()},
\code{\link{util_totals}()}
}
\concept{utility functions}
