% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_k_marginals.R
\name{util_k_marginals}
\alias{util_k_marginals}
\title{Calculate the k-marginals metric}
\usage{
util_k_marginals(eval_data, k = 3)
}
\arguments{
\item{eval_data}{An \code{eval_data} object.}

\item{k}{Scalar order of the k-marginal (valid range = 1:3).}
}
\value{
A value in range \link{0, 1000} where a higher value denotes lower MabsDDs
and consequently greater similarity between confidential and synthetic data.
For multiple replicates, a list of such values, one per replicate.
}
\description{
For each unique k-combination of variables shared by the
synthetic and confidential data, the mean absolute difference (MabsDD)
between the two datasets' marginal cell probabilities is computed. The
MabsDDs are averaged across combinations and rescaled as (1 - mean) * 1000.
}
