% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ks_distance.R
\name{util_ks_distance}
\alias{util_ks_distance}
\title{Calculate the Kolmogorov-Smirnov distance (D) for each numeric variable in
the synthetic and confidential data}
\usage{
util_ks_distance(eval_data, na.rm = FALSE)
}
\arguments{
\item{eval_data}{An \code{eval_data} object}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\value{
A tibble with the D and location of the largest distance for each
numeric variable, one per synthetic data replicate
}
\description{
Calculate the Kolmogorov-Smirnov distance (D) for each numeric variable in
the synthetic and confidential data
}
\seealso{
Other Utility metrics:
\code{\link[=add_pmse]{add_pmse()}},
\code{\link[=add_pmse_ratio]{add_pmse_ratio()}},
\code{\link[=add_propensities]{add_propensities()}},
\code{\link[=add_propensities_tuned]{add_propensities_tuned()}},
\code{\link[=add_specks]{add_specks()}},
\code{\link[=discrimination]{discrimination()}},
\code{\link[=util_ci_overlap]{util_ci_overlap()}},
\code{\link[=util_co_occurrence]{util_co_occurrence()}},
\code{\link[=util_proportions]{util_proportions()}}
}
\concept{Utility metrics}
