% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ks_distance.R
\name{util_ks_distance}
\alias{util_ks_distance}
\title{Calculate the Kolmogorov-Smirnov distance (D) for each numeric variable in
the synthetic and confidential data}
\usage{
util_ks_distance(postsynth, data)
}
\arguments{
\item{postsynth}{A postsynth object or tibble with synthetic data}

\item{data}{A data frame with the original data}
}
\value{
A tibble with the D and location of the largest distance for each
numeric variable
}
\description{
Calculate the Kolmogorov-Smirnov distance (D) for each numeric variable in
the synthetic and confidential data
}
