% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_moments.R
\name{util_moments}
\alias{util_moments}
\title{Calculate summary statistics for original and synthetic data.}
\usage{
util_moments(
  eval_data,
  weight_var = NULL,
  group_by = NULL,
  drop_zeros = FALSE,
  common_vars = TRUE,
  synth_vars = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{eval_data}{An \code{eval_data} object}

\item{weight_var}{An unquoted name of a weight variable}

\item{group_by}{The unquoted name of a (or multiple) grouping variable(s)}

\item{drop_zeros}{A logical for if zeros should be dropped}

\item{common_vars}{A logical for if only common variables should be kept}

\item{synth_vars}{A logical for if \emph{only} synthesized variables should be kept}

\item{na.rm}{A logical for ignoring \code{NA} values in computations.}
}
\value{
A \code{tibble} of summary statistics.
}
\description{
Calculate summary statistics for original and synthetic data.
}
\seealso{
Other utility metrics: 
\code{\link{util_corr_fit}()},
\code{\link{util_na_amount}()},
\code{\link{util_na_cluster}()},
\code{\link{util_percentiles}()},
\code{\link{util_totals}()}
}
\concept{utility metrics}
