% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_moments.R
\name{util_moments}
\alias{util_moments}
\title{Calculate summary statistics for original and synthetic data.}
\usage{
util_moments(
  postsynth,
  data,
  weight_var = 1,
  group_by = NULL,
  drop_zeros = FALSE
)
}
\arguments{
\item{postsynth}{A postsynth object or tibble with synthetic data}

\item{data}{A data frame with the original data}

\item{weight_var}{An unquoted name of a weight variable}

\item{group_by}{The unquoted name of a (or multiple) grouping variable(s)}

\item{drop_zeros}{A Boolean for if zeros should be dropped}
}
\value{
A \code{tibble} of summary statistics.
}
\description{
Calculate summary statistics for original and synthetic data.
}
\seealso{
Other utility metrics: 
\code{\link{util_co_occurrence}()},
\code{\link{util_corr_fit}()},
\code{\link{util_percentiles}()},
\code{\link{util_tails}()},
\code{\link{util_totals}()}
}
\concept{utility metrics}
