% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_na_amount.R
\name{util_na_amount}
\alias{util_na_amount}
\title{Calculate the proportion of missing values in each variable}
\usage{
util_na_amount(eval_data, na_values = NULL)
}
\arguments{
\item{eval_data}{An \code{eval_data} object}

\item{na_values}{An optional scalar or vector of values (in addition to \code{NA})
that should be treated as missing}
}
\value{
A tibble \code{na_prop} (one per synthetic data replicate) with one row
per \code{variable} x \code{source} (\code{"original"}, \code{"synthetic"}, and \code{"holdout"}
when \code{eval_data$holdout_data} is supplied), with the proportion of missing
values in \code{na_prop}.
}
\description{
Calculate the proportion of missing values in each variable
}
\examples{
ed <- eval_data(conf_data = penguins_conf, synth_data = penguins_postsynth)

util_na_amount(ed)

}
\seealso{
Other utility metrics:
\code{\link[=util_corr_fit]{util_corr_fit()}},
\code{\link[=util_moments]{util_moments()}},
\code{\link[=util_percentiles]{util_percentiles()}},
\code{\link[=util_totals]{util_totals()}}
}
\concept{utility metrics}
