% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_na_cluster.R
\name{util_na_cluster}
\alias{util_na_cluster}
\title{Calculate Pearson's linear correlation coefficient clustering of missingness across variables}
\usage{
util_na_cluster(eval_data, na_values = NULL)
}
\arguments{
\item{eval_data}{An \code{eval_data} object}

\item{na_values}{A character vector of values that should be treated as
missing in addition to \code{NA}}
}
\value{
A \code{list} of fit metrics (one per each synthetic data replicate),
restricted to variables with at least one missing value in
\code{eval_data$conf_data} or \code{eval_data$synth_data}:
\itemize{
\item \code{na_cluster_confidential}: phi coefficient matrix (correlation of binary
missingness indicators) of the confidential data.
\item \code{na_cluster_synthetic}: phi coefficient matrix of the synthetic data.
\item \code{na_cluster_holdout}: phi coefficient matrix of the holdout data (if
\code{eval_data$holdout_data} is supplied).
\item \code{na_cluster_difference}: difference between \code{na_cluster_synthetic} and
\code{na_cluster_confidential}.
\item \code{na_cluster_difference_mae}: mean absolute error between
\code{na_cluster_confidential} and \code{na_cluster_synthetic}.
\item \code{na_cluster_difference_rmse}: root mean square error between
\code{na_cluster_confidential} and \code{na_cluster_synthetic}.
}
}
\description{
Calculate Pearson's linear correlation coefficient clustering of missingness across variables
}
\examples{
conf_data <- penguins_conf
conf_data$bill_length_mm[1:10] <- NA
conf_data$bill_depth_mm[1:10] <- NA

ed <- eval_data(conf_data = conf_data, synth_data = penguins_postsynth)

util_na_cluster(ed)

}
\seealso{
Other utility metrics:
\code{\link[=util_corr_fit]{util_corr_fit()}},
\code{\link[=util_moments]{util_moments()}},
\code{\link[=util_na_amount]{util_na_amount()}},
\code{\link[=util_percentiles]{util_percentiles()}},
\code{\link[=util_totals]{util_totals()}}
}
\concept{utility metrics}
