% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_percentiles.R
\name{util_percentiles}
\alias{util_percentiles}
\title{Calculate summary statistics for original and synthetic data.}
\usage{
util_percentiles(
  eval_data,
  probs = c(0.1, 0.5, 0.9),
  weight_var = NULL,
  group_by = NULL,
  drop_zeros = FALSE,
  common_vars = TRUE,
  synth_vars = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{eval_data}{An \code{eval_data} object.}

\item{probs}{A numeric vector of probabilities with values in [0,1]. The
percentiles are interpolated using an empirical CDF. It's possible that the
percentiles are an approximation; especially when weights are used.}

\item{weight_var}{An unquoted name of a weight variable}

\item{group_by}{The unquoted name(s) of a (or multiple) grouping variable(s)}

\item{drop_zeros}{A Boolean for if zeros should be dropped}

\item{common_vars}{A logical for if only common variables should be kept.
This option will frequently result in an error because quantile() is strict
about missing values.}

\item{synth_vars}{A logical for if only synthesized variables should be kept}

\item{na.rm}{A logical for ignoring \code{NA} values in computations.}
}
\value{
A \code{tibble} of summary statistics (one per synthetic data replicate)
}
\description{
Calculate summary statistics for original and synthetic data.
}
\seealso{
Other utility metrics: 
\code{\link{util_corr_fit}()},
\code{\link{util_moments}()},
\code{\link{util_na_amount}()},
\code{\link{util_na_cluster}()},
\code{\link{util_totals}()}
}
\concept{utility metrics}
