% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_percentiles.R
\name{util_percentiles}
\alias{util_percentiles}
\title{Calculate summary statistics for original and synthetic data.}
\usage{
util_percentiles(
  postsynth,
  data,
  probs = c(0.1, 0.5, 0.9),
  weight_var = NULL,
  drop_zeros = FALSE
)
}
\arguments{
\item{postsynth}{A postsynth object or tibble with synthetic data}

\item{data}{A data frame with the original data}

\item{probs}{A numeric vector of probabilities with values in \link{0,1}. The
percentiles are interpolated using an empirical CDF. It's possible that the
percentiles are an approximation; especially when weights are used.}

\item{weight_var}{An unquoted name of a weight variable}

\item{drop_zeros}{A Boolean for if zeros should be dropped}
}
\value{
A \code{tibble} of summary statistics.
}
\description{
Calculate summary statistics for original and synthetic data.
}
\seealso{
Other utility functions: 
\code{\link{util_corr_fit}()},
\code{\link{util_moments}()},
\code{\link{util_tails}()}
}
\concept{utility functions}
