% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_proportions.R
\name{util_proportions}
\alias{util_proportions}
\title{Calculate frequency tables for categorical variables}
\usage{
util_proportions(postsynth, data, group_var = NULL)
}
\arguments{
\item{postsynth}{A postsynth object or tibble with synthetic data}

\item{data}{A data frame with the original data}

\item{group_var}{}
}
\value{
A tibble with variables, classes, and relative frequencies
}
\description{
Calculate frequency tables for categorical variables
}
