% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_proportions.R
\name{util_proportions}
\alias{util_proportions}
\title{Calculate relative frequency tables for categorical variables}
\usage{
util_proportions(
  postsynth,
  data,
  weight_var = NULL,
  group_by = NULL,
  common_vars = TRUE,
  synth_vars = TRUE,
  keep_empty_levels = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{postsynth}{A postsynth object or tibble with synthetic data}

\item{data}{A data frame with the original data}

\item{weight_var}{An unquoted name of a weight variable}

\item{group_by}{An unquoted name of a (or multiple) grouping variable(s)}

\item{common_vars}{A logical for if only common variables should be kept}

\item{synth_vars}{A logical for if only synthesized variables should be kept}

\item{keep_empty_levels}{A logical for keeping all class levels in the group_by
statements, including missing levels.}

\item{na.rm}{A logical for ignoring \code{NA} values in proportion calculations.}
}
\value{
A tibble with variables, classes, and relative frequencies
}
\description{
Calculate relative frequency tables for categorical variables
}
\seealso{
Other Utility metrics: 
\code{\link{add_pmse}()},
\code{\link{add_pmse_ratio}()},
\code{\link{add_propensities}()},
\code{\link{add_propensities_tuned}()},
\code{\link{add_specks}()},
\code{\link{discrimination}()},
\code{\link{util_ci_overlap}()},
\code{\link{util_co_occurrence}()},
\code{\link{util_ks_distance}()}
}
\concept{Utility metrics}
