% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_proportions.R
\name{util_proportions}
\alias{util_proportions}
\title{Calculate relative frequency tables for categorical variables}
\usage{
util_proportions(postsynth, data, weight_var = 1, group_by = NULL)
}
\arguments{
\item{postsynth}{A postsynth object or tibble with synthetic data}

\item{data}{A data frame with the original data}

\item{weight_var}{An unquoted name of a weight variable}

\item{group_by}{An unquoted name of a grouping variable}
}
\value{
A tibble with variables, classes, and relative frequencies
}
\description{
Calculate relative frequency tables for categorical variables
}
