% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tails.R
\name{util_tails}
\alias{util_tails}
\title{Explore the tails of numeric variables}
\usage{
util_tails(postsynth, data, n = 10, weight_var = 1, end = "max")
}
\arguments{
\item{postsynth}{A postsynth object or tibble with synthetic data}

\item{data}{A data frame with the original data}

\item{n}{The number of observations to consider for each variable}

\item{weight_var}{An unquoted name of a weight variable}

\item{end}{"min" for minimum values and "max" for maximum values}
}
\value{
A \code{tibble} of summary statistics.
}
\description{
Explore the tails of numeric variables
}
\seealso{
Other utility metrics: 
\code{\link{util_corr_fit}()},
\code{\link{util_moments}()},
\code{\link{util_percentiles}()},
\code{\link{util_totals}()}
}
\concept{utility metrics}
