% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_totals.R
\name{util_totals}
\alias{util_totals}
\title{Calculate totals for original and synthetic data.}
\usage{
util_totals(
  eval_data,
  weight_var = NULL,
  group_by = NULL,
  common_vars = TRUE,
  synth_vars = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{eval_data}{An \code{eval_data} object}

\item{weight_var}{An unquoted name of a weight variable}

\item{group_by}{The unquoted name of a (or multiple) grouping variable(s)}

\item{common_vars}{A logical for if only common variables should be kept}

\item{synth_vars}{A logical for if only synthesized variables should be kept}

\item{na.rm}{A logical for ignoring \code{NA} values in computations.}
}
\value{
A \code{tibble} of totals.
}
\description{
Calculate totals for original and synthetic data.
}
\seealso{
Other utility metrics: 
\code{\link{util_bivariate}()},
\code{\link{util_corr_fit}()},
\code{\link{util_moments}()},
\code{\link{util_percentiles}()}
}
\concept{utility metrics}
